requirements.txt
streamlit
PyPDF2
google-generativeai>=0.5.0


